add LIDT, LGDT, MOV between control registers
support extended opcodes that have a rmbyte opcode
move check for invalid opcode to after the group/extended table lookups

+++ x86Disasm/Attribute.cs
@@ -4,7 +4,7 @@ namespace x86Disasm
     [System.AttributeUsage(AttributeTargets.Method, AllowMultiple = true)]
     public class CPUFunction : Attribute
     {
-        public ushort OpCode;
+        public uint OpCode;
         public int Count;
 
         public CPUFunction()
